Data Gathering Speed Up
I have managed to make this program 50 times faster after thinking about
the API a bit more. Using the same playlists as the initial data mine, I
have recorded a time of 138 seconds (2 mins 18 seconds) which is
substantially faster than the first mine. This also found 4407 data
points which is 18 more songs than the last mine but that will be
because of people editing their playlists.