Visualiser Data Passed
Created a way for data to be passed to visualiser (requested) and did
some cleanups and bug fixes (songs not being found printing a type
error). This includes a sample file that would be passed to the
visualiser currently. Added a method to get length of a song as
requested by visauls team.