Threading URI Search
Managed to impliment threading for URI search. Had to mute a few things
in spotipy and make things run a bit faster but it shouldn't affect the
rest of the program.
Made a temporary change in the mood detection method that is still empty
to accout for the GUI link to occur soon (Ryan).
Also waiting on a dialog that appears to link pass points to for the
user to understand what is occuring (Ryan).